reoganized auxiliary cart functions
