--- src/utils/cart/calcNumItemsInCart.js
@@ -1,5 +1,0 @@
-export default function calcNumItemsInCart(cart, itemId) {
-  const number = cart.filter((item) => item === itemId).length;
-
-  return number;
-}
