--- src/utils/cart/getCartFromStorage.js
@@ -1,6 +1,0 @@
-export default function getCartFromStorage() {
-  let storedCart = JSON.parse(localStorage.getItem('cart'));
-
-  if (!storedCart) storedCart = [];
-  return storedCart;
-}
