--- src/utils/cart/saveCartToStorage.js
@@ -1,3 +1,0 @@
-export default function saveCartToStorage(itemsList) {
-  localStorage.setItem('cart', JSON.stringify(itemsList));
-}
